uplaod resume and applied candidates model done

+++ backend/job/views.py
@@ -1,17 +1,18 @@
-
 from django.shortcuts import render
-from rest_framework.decorators import api_view
+from django.utils import timezone
+from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
 from rest_framework import status
 from django.db.models import Avg, Min, Max, Count
-
-from .serializers import JobSerializer
-from .models import Job
-from .filters import JobsFilter
-
 from rest_framework.pagination import PageNumberPagination
 
+from rest_framework.permissions import IsAuthenticated
+
+# from .serializers import CandidatesAppliedSerializer, JobSerializer
+from .models import CandidatesApplied, Job
+
 from django.shortcuts import get_object_or_404
+from .filters import JobsFilter
 
 # Create your views here.
 
@@ -48,7 +49,9 @@ def getJob(request, pk):
 
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def newJob(request):
+    request.data['user'] = request.user
     data = request.data
 
     job = Job.objects.create(**data)
@@ -58,8 +61,13 @@ def newJob(request):
 
 
 @api_view(['PUT'])
+@permission_classes([IsAuthenticated])
 def updateJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+    
+    if(job.user != request.user):
+        return Response({'Message': "You are not authorized to perform such this action"}, status = status.HTTP_403_FORBIDDEN)
+    
 
     job.title = request.data['title']
     job.description = request.data['description']
@@ -83,6 +91,10 @@ def updateJob(request, pk):
 @api_view(['DELETE'])
 def deleteJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+    
+    if(job.user != request.user):
+        return Response({'Message': "You are not authorized to perform this action"}, status = status.HTTP_403_FORBIDDEN)
+    
 
     job.delete()
 
@@ -107,3 +119,35 @@ def getTopicStats(request, topic):
     )
 
     return Response(stats)
+
+@api_view(['POST'])
+@permission_classes([IsAuthenticated])
+def applyToJob(request, pk):
+
+    user = request.user
+    job = get_object_or_404(Job, id=pk)
+ 
+    if user.userprofile.resume == '':
+        return Response({ 'error': 'Please upload your resume first' }, status=status.HTTP_400_BAD_REQUEST)
+
+    if job.lastDate < timezone.now():
+        return Response({ 'error': 'You can not apply to this job. Date is over' }, status=status.HTTP_400_BAD_REQUEST)
+
+    alreadyApplied = job.candidatesapplied.filter(user=user).exists()
+
+    if alreadyApplied:
+        return Response({ 'error': 'You have already apply to this job.' }, status=status.HTTP_400_BAD_REQUEST)
+
+
+    jobApplied = CandidatesApplied.objects.create(
+        job = job,
+        user = user,
+        resume = user.userprofile.resume
+    )
+
+    return Response({
+        'applied': True,
+        'job_id': jobApplied.id
+    },
+    status=status.HTTP_200_OK
+    )
